Added task 1 from dz 16
