+++ app/src/main/AndroidManifest.xml
@@ -10,7 +10,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.TMS_Leassons">
         <activity
-            android:name=".MainActivity"
+            android:name=".MainActivityDz16"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
