+++ app/src/main/res/layout/activity_main.xml
@@ -4,13 +4,13 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity2">
+    tools:context=".MainActivityDz16">
 
     <TextView
         android:id="@+id/tx"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Here will be result"
+        android:text="@string/here_will_be_result"
         android:textColor="#9C27B0"
         android:textSize="20sp"
         app:layout_constraintBottom_toBottomOf="parent"
@@ -24,11 +24,12 @@
         android:layout_height="wrap_content"
         android:layout_marginBottom="120dp"
         android:onClick="onclick"
-        android:text="The Batton"
+        android:text="@string/result"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.498"
-        app:layout_constraintStart_toStartOf="parent" />
+        app:layout_constraintStart_toStartOf="parent"
+        tools:ignore="UsingOnClickInXml" />
 
     <TextView
         android:id="@+id/textViewInfo"
@@ -36,7 +37,7 @@
         android:layout_height="wrap_content"
         android:layout_marginStart="8dp"
         android:layout_marginEnd="8dp"
-        android:text="If you push the batton,You will see sum digital from 1 to 100"
+        android:text="2 random numbers, + or * depending on a = Even or not"
         android:textColor="#0A2DF3"
         android:textColorHint="#0927CF"
         android:textSize="20sp"
@@ -44,5 +45,15 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
+
+    <TextView
+        android:id="@+id/pokaz"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:textSize="20sp"
+        app:layout_constraintBottom_toTopOf="@+id/tx"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/textViewInfo" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
